+++ src/MyMath/Vector/Vector3.java
@@ -2,9 +2,9 @@ package MyMath.Vector;
 
 public class Vector3
 {
-    private float x;
-    private float y;
-    private float z;
+    private double x;
+    private double y;
+    private double z;
 
     public Vector3()
     {
@@ -17,7 +17,7 @@ public class Vector3
         System.out.println("X: " + this.x + " Y: " + this.y + " Z: " + this.z);
     }
 
-    public Vector3(float x, float y, float z)
+    public Vector3(double x, double y, double z)
     {
         this.x = x;
         this.y = y;
@@ -26,55 +26,55 @@ public class Vector3
 
     public Vector3 add(Vector3 b)
     {
-        float x = this.x + b.x;
-        float y = this.y + b.y;
-        float z = this.z + b.z;
+        double x = this.x + b.x;
+        double y = this.y + b.y;
+        double z = this.z + b.z;
         return new Vector3(x, y, z);
     }
 
     public Vector3 sub(Vector3 b)
     {
-        float x = this.x - b.x;
-        float y = this.y - b.y;
-        float z = this.z - b.z;
+        double x = this.x - b.x;
+        double y = this.y - b.y;
+        double z = this.z - b.z;
         return new Vector3(x, y, z);
     }
 
-    public Vector3 mul(float b)
+    public Vector3 mul(double b)
     {
-        float x = this.x * b;
-        float y = this.y * b;
-        float z = this.z * b;
+        double x = this.x * b;
+        double y = this.y * b;
+        double z = this.z * b;
         return new Vector3(x, y, z);
     }
 
     public Vector3 mul(Vector3 b)
     {
-        float x = this.x * b.x;
-        float y = this.y * b.x;
-        float z = this.z * b.x;
+        double x = this.x * b.x;
+        double y = this.y * b.x;
+        double z = this.z * b.x;
         return new Vector3(x, y, z);
     }
 
-    public Vector3 div(float b)
+    public Vector3 div(double b)
     {
-        float x = this.x / b;
-        float y = this.y / b;
-        float z = this.z / b;
+        double x = this.x / b;
+        double y = this.y / b;
+        double z = this.z / b;
         return new Vector3(x, y, z);
     }
 
-    public float magnitudeSquared()
+    public double magnitudeSquared()
     {
-        float x = this.x;
-        float y = this.y;
-        float z = this.z;
+        double x = this.x;
+        double y = this.y;
+        double z = this.z;
         return x*x+y*y+z*z;
     }
 
-    public float magnitude()
+    public double magnitude()
     {
-        return (float) Math.sqrt(this.magnitudeSquared());
+        return (double) Math.sqrt(this.magnitudeSquared());
     }
 
     public Vector3 normalize()
@@ -82,40 +82,40 @@ public class Vector3
         return this.div(this.magnitude());
     }
 
-    public float dot(Vector3 b)
+    public double dot(Vector3 b)
     {
-        float x = this.x * b.x;
-        float y = this.y * b.y;
-        float z = this.z * b.z;
+        double x = this.x * b.x;
+        double y = this.y * b.y;
+        double z = this.z * b.z;
         return x + y + z;
     }
 
-    public float getX()
+    public double getX()
     {
         return x;
     }
 
-    public void setX(float x)
+    public void setX(double x)
     {
         this.x = x;
     }
 
-    public float getY()
+    public double getY()
     {
         return y;
     }
 
-    public void setY(float y)
+    public void setY(double y)
     {
         this.y = y;
     }
 
-    public float getZ()
+    public double getZ()
     {
         return z;
     }
 
-    public void setZ(float z)
+    public void setZ(double z)
     {
         this.z = z;
     }
